Add a fix for mangling longer CHANGELOGs
This reads the whole file in and closes it before attempting to open
the file for writing, thereby avoiding the situation where we start
writing to the file (and truncating it) before we've read it in, or
overwrite the existing file and leave old bytes at the end.

There are no tests for this because the code in root.go is not (yet!)
tested. I've tested it manually.

Co-authored-by: Carl Menezes <carleeto@gmail.com>